+++ lib/view_component_contrib/railtie.rb
@@ -2,11 +2,25 @@
 
 module ViewComponentContrib
   class Railtie < Rails::Railtie
-    config.view_component.preview_paths << File.join(ViewComponentContrib::APP_PATH, "views")
+    if !config.view_component.previews.nil?
+      config.view_component.previews.paths << File.join(ViewComponentContrib::APP_PATH, "views")
+    else
+      config.view_component.preview_paths << File.join(ViewComponentContrib::APP_PATH, "views")
+    end
 
     initializer "view_component-contrib.skip_loading_previews_if_disabled" do
-      unless Rails.application.config.view_component.show_previews
-        previews = Rails.application.config.view_component.preview_paths.flat_map do |path|
+      vc_config = Rails.application.config.view_component
+
+      preview_enabled = (!vc_config.previews.nil?) ?
+                          vc_config.previews.enabled :
+                          vc_config.show_previews
+
+      unless preview_enabled
+        preview_paths = (!vc_config.previews.nil?) ?
+                          vc_config.previews.paths :
+                          vc_config.preview_paths
+
+        previews = preview_paths.flat_map do |path|
           Pathname(path).glob("**/*preview.rb")
         end
         Rails.autoloaders.each { |autoloader| autoloader.ignore(previews) }