+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- Fix compatibility with view_component v4. 
+
 ## 0.2.4 (2025-01-03)
 
 - Add inheritance strategies to style variants ([@omarluq][])